Rename feature folder to `action-menu`. Update `elements` readme with link to ActionMenu.md.
