+++ packages/elements/src/index.ts
@@ -13,10 +13,10 @@ export * from "./components/ui/badge/Badge";
 export * from "./components/ui/badge/WithBadge";
 export * from "./components/ui/toggle-button/ToggleButtonTheme";
 export * from "./components/ui/toggle-button/ToggleButton";
-export * from "./components/ui/action-dropdown/ActionDropdown";
-export * from "./components/ui/action-dropdown/ActionDropdownTheme";
-export * from "./components/ui/action-dropdown/ActionMenu";
-export * from "./components/ui/action-dropdown/ActionMenuContext";
-export * from "./components/ui/action-dropdown/ActionMenuItem";
-export * from "./components/ui/action-dropdown/ActionMenuSeparator";
-export * from "./components/ui/action-dropdown/ActionMenuTheme";
+export * from "./components/ui/action-menu/ActionDropdown";
+export * from "./components/ui/action-menu/ActionDropdownTheme";
+export * from "./components/ui/action-menu/ActionMenu";
+export * from "./components/ui/action-menu/ActionMenuContext";
+export * from "./components/ui/action-menu/ActionMenuItem";
+export * from "./components/ui/action-menu/ActionMenuSeparator";
+export * from "./components/ui/action-menu/ActionMenuTheme";
